Fix argument type error in lead_detail_screen.dart

+++ lib/features/leads/presentation/lead_detail_screen.dart
@@ -273,7 +273,9 @@ class _LeadDetailScreenState extends State<LeadDetailScreen> {
                   final horizontalPadding =
                       (ResponsiveUtils.isDesktop(context) ||
                               ResponsiveUtils.isWide(context))
-                          ? (baseHorizontalPadding - 4).clamp(16, double.infinity)
+                          ? (baseHorizontalPadding - 4)
+                              .clamp(16.0, double.infinity)
+                              .toDouble()
                           : baseHorizontalPadding;
                   final baseContentMaxWidth =
                       ResponsiveUtils.contentMaxWidth(context);
